Moves Batch Converter Code
Separates batch converter code into batch_converter.py
- Additionally adds overwrite confirmation check
